make transform panel a shared component
