+++ src/plugins/explore/public/components/data_transformations/transformations/convert_field_type_transformation.tsx
@@ -9,7 +9,7 @@ import { EuiButtonIcon, EuiFlexGroup, EuiFlexItem, EuiSelect } from '@elastic/eu
 import { i18n } from '@osd/i18n';
 import { TransformationInstance, TransformationDefinition, FieldSchema } from '../index';
 import { FieldSelector } from '../field_selector';
-import { OpenSearchSearchHit } from '../../../../types/doc_views_types';
+import { OpenSearchSearchHit } from '../../../types/doc_views_types';
 
 type TargetType = 'string' | 'number' | 'boolean' | 'date';
